新增注册账号和找回密码功能

+++ hospitalFront/src/App.vue
@@ -15,6 +15,9 @@ export default {
       // 监听路由变化，防止浏览器输入地址，没登录就直接进入页面
       $route: {
         handler: function(val, oldVal){
+            if (val.name === 'register' || val.name === 'forgetPassword') {
+                return
+            }
             if (val.name === 'login') {
                 sessionStorage.clear()
             } else {
